refactor: spostamento metodi in Utils
